api文档优化  #8594
