+++ src/backend/ci/core/openapi/biz-openapi/src/main/resources/i18n/message_en_US.properties
@@ -25,7 +25,7 @@ bkPolymorphicClassImplementation=Polymorphic class implementation
 bkPolymorphismModelIllustrate=*At the implementation of the polymorphic base class <{0}>, where the field {1} = [{2}] is specified as the class implementation*
 bkReturnsTheSample=Returns sample
 bkErrorPrompt=**Note: Make sure that Content-Type: applicationjson is present in the header, otherwise the request returns a 415 error code**
-bkCurlPrompt='[Replace it with the API address bar above for the request address]{0}'
+bkCurlPrompt=[Replace it with the API address bar above for the request address]{0}
 bkPayloadRequestSample=< {0} >, Then the request should be:
 bkTheFieldIsReadOnly=The field is read-only
 bkObjectPropertyIllustrate=Any type, referring to the actual request or return
