+++ src/backend/ci/core/openapi/biz-openapi/src/main/resources/i18n/message_zh_CN.properties
@@ -25,7 +25,7 @@ bkPolymorphicClassImplementation=多态类实现
 bkPolymorphismModelIllustrate=*多态基类 <{0}> 的实现处, 其中当字段 {1} = [{2}] 时指定为该类实现*
 bkReturnsTheSample=返回样例
 bkErrorPrompt=**注意: 确保 header 中存在 Content-Type: application/json ,否则请求返回415错误码**
-bkCurlPrompt='[请替换为上方API地址栏请求地址]{0}'
+bkCurlPrompt=[请替换为上方API地址栏请求地址]{0}
 bkPayloadRequestSample=< {0} >, 那么请求应该为:
 bkTheFieldIsReadOnly=该字段只读
 bkObjectPropertyIllustrate=Any 任意类型，参照实际请求或返回
